+++ src/index.ts
@@ -29,16 +29,27 @@ const logger = logging.getLogger('polymer-project-config');
 export const defaultSourceGlobs = ['src/**/*'];
 
 /**
- * Resolve any glob to the given path, even if glob
+ * Resolve any glob or path from the given path, even if glob
  * is negative (begins with '!').
  */
-function resolveGlob(fromPath: string, glob: string): string {
+function globResolve(fromPath: string, glob: string): string {
   if (glob.startsWith('!')) {
     const includeGlob = glob.substring(1);
     return '!' + path.resolve(fromPath, includeGlob);
   } else {
     return path.resolve(fromPath, glob);
   }
+}
+
+/**
+ * Returns a relative path for a glob or path, even if glob
+ * is negative (begins with '!').
+ */
+function globRelative(fromPath: string, glob: string): string {
+  if (glob.startsWith('!')) {
+    return '!' + path.relative(fromPath, glob.substr(1));
+  }
+  return path.relative(fromPath, glob);
 }
 
 /**
@@ -158,7 +169,7 @@ export class ProjectConfig {
    *
    * TODO: make this method and the one below async.
    */
-  static loadOptionsFromFile(filepath: string): ProjectOptions {
+  static loadOptionsFromFile(filepath: string): ProjectOptions|null {
     try {
       const configContent = fs.readFileSync(filepath, 'utf-8');
       const contents = JSON.parse(configContent);
@@ -186,7 +197,7 @@ export class ProjectConfig {
    * Given an absolute file path to a polymer.json-like ProjectOptions object,
    * return a new ProjectConfig instance created with those options.
    */
-  static loadConfigFromFile(filepath: string): ProjectConfig {
+  static loadConfigFromFile(filepath: string): ProjectConfig|null {
     let configParsed = ProjectConfig.loadOptionsFromFile(filepath);
     if (!configParsed) {
       return null;
@@ -240,13 +251,13 @@ export class ProjectConfig {
      * extraDependencies
      */
     this.extraDependencies = (options.extraDependencies ||
-                              []).map((glob) => resolveGlob(this.root, glob));
+                              []).map((glob) => globResolve(this.root, glob));
 
     /**
      * sources
      */
     this.sources = (options.sources || defaultSourceGlobs)
-                       .map((glob) => resolveGlob(this.root, glob));
+                       .map((glob) => globResolve(this.root, glob));
     this.sources.push(this.entrypoint);
     if (this.shell) {
       this.sources.push(this.shell);
@@ -350,22 +361,45 @@ export class ProjectConfig {
           const buildPreset = build.preset;
           console.assert(
               !buildPreset || isValidPreset(buildPreset),
-              `${validateErrorPrefix}: "${buildPreset}" is not a valid `
-              + ` "builds" preset.`);
+              `${validateErrorPrefix}: "${buildPreset}" is not a valid ` +
+                  ` "builds" preset.`);
           console.assert(
               buildName,
               `${validateErrorPrefix}: all "builds" require ` +
                   `a "name" property when there are multiple builds defined.`);
           console.assert(
-              !buildNames.has(buildName),
+              !buildNames.has(buildName!),
               `${validateErrorPrefix}: "builds" duplicate build name ` +
                   `"${buildName}" found. Build names must be unique.`);
-          buildNames.add(buildName);
+          buildNames.add(buildName!);
         }
       }
     }
 
     return true;
+  }
+
+  /**
+   * Generate a JSON string serialization of this configuration. File paths
+   * will be relative to root.
+   */
+  toJSON(): string {
+    const obj = {
+      entrypoint: globRelative(this.root, this.entrypoint),
+      shell: this.shell ? globRelative(this.root, this.shell) : undefined,
+      fragments: this.fragments.map((absolutePath) => {
+        return globRelative(this.root, absolutePath);
+      }),
+      sources: this.sources.map((absolutePath) => {
+        return globRelative(this.root, absolutePath);
+      }),
+      extraDependencies: this.extraDependencies.map((absolutePath) => {
+        return globRelative(this.root, absolutePath);
+      }),
+      builds: this.builds,
+      lint: this.lint,
+    };
+    return JSON.stringify(obj, null, 2);
   }
 }
 
@@ -373,7 +407,7 @@ export class ProjectConfig {
 // interface for runtime validation. See the build script in package.json for
 // more info.
 const getSchema: () => jsonschema.Schema = (() => {
-  let schema: jsonschema.Schema|undefined = undefined;
+  let schema: jsonschema.Schema;
 
   return () => {
     if (schema === undefined) {