+++ CHANGELOG.md
@@ -5,11 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
-
+## Unreleased
+* `lint`: Add support for `minSeverity` build option that controls the severity of lint warnings to output and trigger failure.
 <!-- Add new, unreleased changes here. -->
 
-* `lint`: Add support for `minSeverity` build option that controls the severity of lint warnings to output and trigger failure.
+## [3.4.0] - 2017-06-21
+* Modified the `bundle` property in project build options to support the subset of `polymer-bundler` options which can be serialized in a JSON file.
+
+## [3.3.0] - 2017-06-09
+* Removed `insertPrefetchLinks` from build presets. See https://github.com/Polymer/polymer-build/issues/239 for details.
+
+## [3.2.1] - 2017-05-19
+* Turned on insertPrefetchLinks for the bundled presets after prefetch link bugs were fixed.
+
+## [3.2.0] - 2017-05-15
+* Add ProjectConfig toJSON method.
+
+## [3.1.0] - 2017-05-15
+* Add browserCapabilities and basePath options.
 
 ## [3.0.0] - 2017-05-09
 * Add build preset support.